refactor: reuse PhotoGrid component in UserLikes and UserPhotos
- Updated UserLikes and UserPhotos to use the reusable PhotoGrid component
- Passed imageExtractor to handle photo URL extraction
- Improved code consistency and maintainability by reducing duplication
